moved SetPackageAction flags to own package to avoid import cycle
